完成动态列表对接

+++ todokotlin/src/main/java/com/union/bangbang/todokotlin/base/data/net/Api.kt
@@ -33,6 +33,6 @@ interface Api {
     fun addMemo(@Body memo: Memo): Observable<InfoEntity<Memo>>
 
     @GET("memo/findMemoByLocation")
-    fun findMemoByLocation(@Query("longitude") longitude: Float, @Query("latitude") latitude: Float, @Query("distance") distance: Float): Observable<InfoEntity<List<Memo>>>
+    fun findMemoByLocation(@Query("longitude") longitude: Double, @Query("latitude") latitude: Double, @Query("distance") distance: Double): Observable<InfoEntity<List<Memo>>>
 
 }
